Move hookpoints to their own file
